Move all questions to prompts file for easy roll-up into parent generator.
